+++ tests/test_utils.py
@@ -1,3 +1,5 @@
+from scim2_models.rfc7643.enterprise_user import EnterpriseUser
+from scim2_models.rfc7643.user import User
 from scim2_models.utils import to_camel
 
 
@@ -13,3 +15,15 @@ def test_to_camel():
     assert to_camel("Foo_Bar") == "fooBar"
 
     assert to_camel("$foo$") == "$foo$"
+
+
+def test_get_extension_for_schema():
+    """Test get_extension_model method on Resource."""
+    user = User[EnterpriseUser]()
+    extension_class = user.get_extension_model(
+        "urn:ietf:params:scim:schemas:extension:enterprise:2.0:User"
+    )
+    assert extension_class == EnterpriseUser
+
+    extension_class = user.get_extension_model("urn:unknown:schema")
+    assert extension_class is None